Fix warnings.

+++ environment/snake/src/core/scenes/game/messages/new_highscore_message.c
@@ -49,7 +49,7 @@ void new_highscore_message_draw_and_wait(int score)
     }
     while(name[0] == '\n');
     
-    for(int i = 0; i < strlen(name); i++)
+    for(size_t i = 0; i < strlen(name); i++)
     {
         if(name[i] == '\n')
         {
